requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-environ==0.4.5
django-extensions==2.2.2
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.9.0
djoser==2.0.3
gcloud==0.17.0
googleapis-common-protos==1.6.0
gunicorn==19.7.1
httplib2==0.14.0
idna==2.8
jws==0.1.3
oauth2client==3.0.0
protobuf==3.10.0rc1
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dotenv==0.10.3
python-jwt==2.0.1
pytz==2019.2
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==3.3.1
quandl
numpy
matplotlib
pandas
sklearn
seaborn
yfinance
python-craigslist